+++ src/components/charts/eurobonus/YearlySpentChart.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Bar } from "react-chartjs-2";
 import { Chart as ChartJS, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend } from "chart.js";
-import type { YearlyTransaction } from "../../models/transaction";
+import type { YearlyTransaction } from "../../../models/transaction";
 import styles from "./YearlySpentChart.module.scss";
 
 ChartJS.register(CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend);