+++ src/utils/viatrumf-calculations.tsx
@@ -8,7 +8,7 @@ export const calculateViatrumfVendorTransactions = (transactions: ViatrumfTransa
     year: new Date(transaction.purchase_date).getFullYear(),
     value: transaction.trumf_bonus ?? 0,
     status: transaction.status,
-    profile_id: transaction.profile_id,
+    profile_id: transaction.profile_id!,
   }));
 
 export const getTrumfValue = (value: number, currency: TrumfCurrency): number => {
@@ -20,4 +20,25 @@ export const getTrumfValue = (value: number, currency: TrumfCurrency): number =>
     default:
       return value;
   }
+};
+
+export const getLatestViatrumfDate = (transactions: ViatrumfTransaction[]): string | null => {
+  if (transactions.length === 0) {
+    return null;
+  }
+  return new Date(Math.max(...transactions.map(transaction => new Date(transaction.purchase_date).getTime()))).toLocaleDateString("en-US", {
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+  });
+};
+
+export const getEarliestViatrumfDate = (transactions: ViatrumfTransaction[]): string | null => {
+  if (transactions.length === 0) {
+    return null;
+  }
+  return new Date(Math.min(...transactions.map(transaction => new Date(transaction.purchase_date).getTime()))).toLocaleDateString("en-US", {
+    year: "numeric",
+    month: "long",
+  });
 };